Add minimal DiffBot example
